remove some queries to prevent overloading api calls
