+++ queries/gov/useProposalsQuery.tsx
@@ -5,7 +5,7 @@ import QUERY_KEYS from 'constants/queryKeys';
 import { SPACE_KEY, snapshotEndpoint } from 'constants/snapshot';
 
 import { appReadyState } from 'store/app';
-import { Proposal, SpaceData, Vote } from './types';
+import { Proposal, SpaceData } from './types';
 import { networkState, walletAddressState } from 'store/wallet';
 import Connector from 'containers/Connector';
 import { ethers } from 'ethers';
@@ -93,68 +93,8 @@ const useProposalsQuery = (spaceKey: SPACE_KEY, options?: QueryConfig<Proposal[]
 
 			const mappedProposals = proposals.map(async (proposal) => {
 				if (validHashes.includes(proposal.id)) {
-					const block = parseInt(proposal.snapshot);
-					const currentBlock = provider?.getBlockNumber() ?? 0;
-					const blockTag = block > currentBlock ? 'latest' : block;
-
-					const { votes }: { votes: Vote[] } = await request(
-						snapshotEndpoint,
-						gql`
-							query Votes($proposal: String) {
-								votes(first: 1000, where: { proposal: $proposal }) {
-									id
-									voter
-									choice
-								}
-							}
-						`,
-						{ proposal: proposal.id }
-					);
-
-					const voterAddresses = votes.map((e: Vote) => ethers.utils.getAddress(e.voter));
-
-					const {
-						scores: { scores },
-					} = await request(
-						snapshotEndpoint,
-						gql`
-							query Scores(
-								$spaceKey: String
-								$strategies: [Any]!
-								$network: String!
-								$addresses: [String]!
-								$snapshot: Any
-							) {
-								scores(
-									space: $spaceKey
-									strategies: $strategies
-									network: $network
-									addresses: $addresses
-									snapshot: $snapshot
-								) {
-									scores
-								}
-							}
-						`,
-						{
-							spaceKey,
-							strategies: space.strategies,
-							network: space.network,
-							addresses: voterAddresses,
-							snapshot: blockTag,
-						}
-					);
-
-					let voteCount = 0;
-
-					space.strategies.forEach((_, i: number) => {
-						let arrayOfVotes = Object.values(scores[i]) as number[];
-						voteCount = voteCount + arrayOfVotes.filter((score: number) => score > 0).length;
-					});
-
 					return {
 						...proposal,
-						votes: voteCount,
 					};
 				} else {
 					return null;
