+++ sections/gov/components/Proposal/Results.tsx
@@ -4,7 +4,6 @@ import { Svg } from 'react-optimized-image';
 import Spinner from 'assets/svg/app/loader.svg';
 import useActiveTab from '../../hooks/useActiveTab';
 import { FlexDivRowCentered } from 'styles/common';
-import { ProposalResults } from 'queries/gov/useProposalQuery';
 import { formatNumber, formatPercent } from 'utils/formatters/number';
 import ProgressBar from 'components/ProgressBar';
 import { MaxHeightColumn, StyledTooltip } from 'sections/gov/components/common';
@@ -13,6 +12,7 @@ import CouncilNominations from 'constants/nominations.json';
 import { useRecoilValue } from 'recoil';
 import { numOfCouncilSeatsState } from 'store/gov';
 import { QueryResult } from 'react-query';
+import { ProposalResults } from 'queries/gov/types';
 
 type ResultsProps = {
 	proposalResults: QueryResult<ProposalResults, unknown>;
